[rust] fix test races and null builds
The depfile is now moved to a new location such that it is only updated when
the specific target is rebuilt, rather than being marked dirty because of
shared dependencies.

Tests are built at the same time as libraries/binaries and their links are
only updated when necessary.

+++ rust/build_target.py
@@ -11,9 +11,16 @@ import string
 import subprocess
 import sys
 
-ROOT_PATH = os.path.abspath(__file__ + "/../../..")
-sys.path += [os.path.join(ROOT_PATH, "third_party", "pytoml")]
-import pytoml
+# Fixes the target path in the given depfile.
+def fix_depfile(depfile_path, output_path, base_path):
+    with open(depfile_path, "r+") as depfile:
+        with open(output_path, "w+") as output_file:
+            content = depfile.read()
+            content_split = content.split(': ', 1)
+            target_path = content_split[0]
+            adjusted_target_path = os.path.relpath(target_path, start=base_path)
+            new_content = "%s: %s" % (adjusted_target_path, content_split[1])
+            output_file.write(new_content)
 
 # Creates the directory containing the given file.
 def create_base_directory(file):
@@ -79,14 +86,8 @@ def main():
                         help="Pass --frozen to cargo when building",
                         default = False,
                         action="store_true")
-    parser.add_argument("--label",
-                        help="Label of the target to build",
-                        required=True)
     parser.add_argument("--cmake-dir",
                         help="Path to the directory containing cmake",
-                        required=True)
-    parser.add_argument("--vendor-directory",
-                        help="Path to the vendored crates",
                         required=True)
     parser.add_argument("--shared-libs-root",
                         help="Path to the location of shared libraries",
@@ -115,14 +116,8 @@ def main():
     env["RANLIB"] = args.clang_prefix + '/llvm-ranlib'
     env["PATH"] = "%s:%s" % (env["PATH"], args.cmake_dir)
 
-    # Generate Cargo.toml.
-    original_manifest = os.path.join(args.crate_root, "Cargo.toml")
     target_directory = os.path.join(args.gen_dir, "target")
     create_base_directory(target_directory)
-    package_name = None
-    with open(original_manifest, "r") as manifest:
-        config = pytoml.load(manifest)
-        package_name = config["package"]["name"]
 
     call_args = [
         args.cargo,
@@ -139,6 +134,11 @@ def main():
         call_args.append("--lib")
     if args.type == "bin":
         call_args.extend(["--bin", args.name])
+
+    if args.with_tests:
+        call_args.append("--tests")
+        call_args.append("--message-format=json")
+
     retcode, stdout, stderr = run_command(call_args, env, args.crate_root)
     if retcode != 0:
         print(stdout + stderr)
@@ -147,49 +147,45 @@ def main():
     # Fix the depfile manually until a flag gets added to cargo to tweak the
     # base path for targets.
     # Note: out_dir already contains the "target.rust" directory.
+    # Note: the depfile is written to a new path, as cargo re-writes depfiles
+    # for all dependencies of a target on every build, which means that a bin
+    # build might "update" an already built depfile and corrupt the paths
+    # contained within.
     output_name = args.name
     if args.type == "lib":
         output_name = "lib%s" % args.name
     build_type = "release" if args.release else "debug"
     depfile_path = os.path.join(args.out_dir, args.target_triple, build_type,
                                 "%s.d" % output_name)
+    fix_depfile(depfile_path, os.path.join(args.out_dir, os.path.basename(depfile_path)), args.root_out_dir)
 
-    if args.with_tests:
-        test_args = list(call_args)
-        test_args[1] = "build"
-        test_args.append("--tests")
-        test_args.append("--message-format=json")
-        retcode, stdout, _ = run_command(test_args, env, args.crate_root)
-        if retcode != 0:
-            # The output is not particularly useful as it is formatted in JSON.
-            # Re-run the command with a user-friendly format instead.
-            del test_args[-1]
-            _, stdout, stderr = run_command(test_args, env, args.crate_root)
-            print(stdout + stderr)
-            return retcode
+    for line in stdout.splitlines():
+        data = json.loads(line)
+
+        if not args.with_tests:
+            continue
         generated_test_path = None
         test_name = None
-        for line in stdout.splitlines():
-            data = json.loads(line)
-            if "profile" in data and data["profile"]["test"]:
-              generated_test_path = data["filenames"][0]
-              # only integration tests are marked as 'test', unit tests are 'lib'
-              if data["target"]["kind"][0] == "test":
-                test_type = 'integration'
-                if "target" in data and data["target"]["name"]:
-                    test_name = data["target"]["name"]
-              else:
-                test_type = 'unit'
-            if generated_test_path and test_type:
-                if test_name:
-                    dest_test_path = os.path.join(args.out_dir,
-                            "%s-%s-%s-%s-test" % (args.name, test_name, args.type, test_type))
-                else:
-                    # maintain support for unit tests
-                    dest_test_path = os.path.join(args.out_dir,
-                            "%s-%s-%s-test" % (args.name, args.type, test_type))
-                if os.path.islink(dest_test_path):
-                    os.unlink(dest_test_path)
+        if not ("profile" in data and data["profile"]["test"]):
+            continue
+        generated_test_path = data["filenames"][0]
+        # only integration tests are marked as 'test', unit tests are 'lib'
+        test_type = 'unit'
+        if data["target"]["kind"][0] == "test":
+            test_type = 'integration'
+            if "target" in data and data["target"]["name"]:
+                test_name = data["target"]["name"]
+        if generated_test_path and test_type:
+            if test_name:
+                dest_test_path = os.path.join(args.out_dir,
+                        "%s-%s-%s-%s-test" % (args.name, test_name, args.type, test_type))
+            else:
+                # maintain support for unit tests
+                dest_test_path = os.path.join(args.out_dir,
+                        "%s-%s-%s-test" % (args.name, args.type, test_type))
+            if os.path.islink(dest_test_path) and os.readlink(dest_test_path) != generated_test_path:
+                os.unlink(dest_test_path)
+            if not os.path.exists(dest_test_path):
                 os.symlink(generated_test_path, dest_test_path)
 
     return 0
